Force organize by row on new tables
If not specifed, DB2 will attempt to create new table based on the
default organize by setting. If it set to column, using ALTER statement
won't be possible. This will make even the default django migration
fail (see auth.0002_alter_permission_name_max_length).
With an explict ORGANIZE BY at the and of the